+++ engine/launcher/src/test/resources/org/enso/launcher/components/fake-releases/enso/enso-0.0.1-pre/manifest.yaml
